+++ src/components/entry/EntryDetail.tsx
@@ -248,8 +248,8 @@ const EntryDetail = ({ triggerReload }: IEntryDetailParams) => {
                 {t('entry.detail.summary')}
               </span>
             </div>
-            <span className={'text-white text-left'}>
-              {entry.response.summary}
+            <span className={'text-white text-left w-full'}>
+              {entry.response.summary ? entry.response.summary : '-'}
             </span>
           </div>
         </div>