+++ src/pages/admin/entries-add/components/SecondStep.tsx
@@ -1,10 +1,12 @@
-import { ChangeEvent, FormEvent } from 'react';
+import { ChangeEvent, FormEvent, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import ElviraInput from '../../../../components/common/ElviraInput';
 import ElviraTextarea from '../../../../components/common/ElviraTextarea';
 import PreviousButton from './PreviousButton';
 import NextButton from './NextButton';
 import { IWizardParams } from '../../../../utils/interfaces/general/general';
+import useCustomEffect from '../../../../hooks/useCustomEffect';
+import ElviraSelect from '../../../../components/common/ElviraSelect';
 
 const SecondStep = ({
   entryForm,
@@ -13,6 +15,58 @@ const SecondStep = ({
   setStepIndex,
 }: IWizardParams) => {
   const { t } = useTranslation();
+
+  const year = new Date().getFullYear();
+
+  const [selectedYear, setSelectedYear] = useState<string>('YYYY');
+  const [selectedMonth, setSelectedMonth] = useState<string>('MM');
+  const [selectedDay, setSelectedDay] = useState<string>('DD');
+  const [maxDay, setMaxDay] = useState<number>(31);
+
+  // when month change change possible day (30;31;28/29)
+  useCustomEffect(() => {
+    if (selectedMonth !== 'MM') {
+      if (['1', '3', '5', '7', '8', '10', '12'].includes(selectedMonth)) {
+        setMaxDay(31);
+      } else if (selectedMonth == '2') {
+        // when its february
+        if (selectedYear !== 'YYYY') {
+          const numYear = parseInt(selectedYear);
+          if (
+            (numYear % 4 === 0 && numYear % 100 !== 0) ||
+            numYear % 400 === 0
+          ) {
+            setMaxDay(29); // Leap year, February has 29 days
+          } else {
+            setMaxDay(28); // Non-leap year, February has 28 days
+          }
+        }
+      } else setMaxDay(30);
+    }
+  }, [selectedMonth, selectedYear]);
+
+  useCustomEffect(() => {
+    if (
+      selectedYear !== 'YYYY' &&
+      selectedMonth !== 'MM' &&
+      selectedDay !== 'DD'
+    ) {
+      setEntryForm({
+        ...entryForm,
+        published_at: `${selectedYear}-${selectedMonth}-${selectedDay}`,
+      });
+    } else if (selectedYear !== 'YYYY' && selectedMonth !== 'MM') {
+      setEntryForm({
+        ...entryForm,
+        published_at: `${selectedYear}-${selectedMonth}`,
+      });
+    } else if (selectedYear !== 'YYYY') {
+      setEntryForm({
+        ...entryForm,
+        published_at: `${selectedYear}`,
+      });
+    }
+  }, [selectedDay, selectedMonth, selectedYear]);
 
   const handleSubmitStepTwo = (e: FormEvent<HTMLFormElement>) => {
     e.preventDefault();
@@ -23,12 +77,6 @@ const SecondStep = ({
     setEntryForm({
       ...entryForm, // Preserve existing properties of entryForm
       title: event.target.value, // Update the title property
-    });
-  };
-  const handleYearChange = (event: ChangeEvent<HTMLInputElement>) => {
-    setEntryForm({
-      ...entryForm, // Preserve existing properties of entryForm
-      published_at: event.target.value, // Update the published_at property
     });
   };
   const handlePublisherChange = (event: ChangeEvent<HTMLInputElement>) => {
@@ -59,11 +107,57 @@ const SecondStep = ({
         placeholder={t('entry.wizard.title')}
         value={entryForm.title}
       />
-      <ElviraInput
-        onChange={handleYearChange}
-        placeholder={t('entry.wizard.year')}
-        value={entryForm.published_at}
-      />
+      <div className='flex flex-col gap-4 w-full'>
+        <span>{t('entry.wizard.year')}</span>
+        <div className='w-fit flex gap-4'>
+          <ElviraSelect
+            name='date-year'
+            value={selectedYear}
+            onChange={(e) => setSelectedYear(e.target.value)}
+          >
+            <option value='YYYY'>YYYY</option>
+            {Array.from({
+              length: year - 1900,
+            }).map((_, i) => (
+              <option key={i} value={(year - i).toString()}>
+                {year - i}
+              </option>
+            ))}
+          </ElviraSelect>
+          {selectedYear !== 'YYYY' && (
+            <ElviraSelect
+              name='date-month'
+              value={selectedMonth}
+              onChange={(e) => setSelectedMonth(e.target.value)}
+            >
+              <option value='MM'>MM</option>
+              {Array.from({
+                length: 12,
+              }).map((_, i) => (
+                <option key={i} value={(i + 1).toString()}>
+                  {i + 1}
+                </option>
+              ))}
+            </ElviraSelect>
+          )}
+          {selectedMonth !== 'MM' && selectedYear !== 'YYYY' && (
+            <ElviraSelect
+              name='date-day'
+              value={selectedDay}
+              onChange={(e) => setSelectedDay(e.target.value)}
+            >
+              <option value='DD'>DD</option>
+              {Array.from({
+                length: maxDay,
+              }).map((_, i) => (
+                <option key={i} value={(i + 1).toString()}>
+                  {i + 1}
+                </option>
+              ))}
+            </ElviraSelect>
+          )}
+        </div>
+      </div>
       <ElviraInput
         onChange={handlePublisherChange}
         placeholder={t('entry.wizard.publisher')}