+++ CHANGELOG.md
@@ -13,10 +13,12 @@
 - **Added**: Scripts for analysis
 - **Added**: Layout for entries
 - **Added**: Catalogs
+- **Added**: `ElviraSelect`
 - **Changed**: Design of `ElviraInput`
 - **Changed**: Design of `ElviraTextarea`
 - **Changed**: In breadcrumbs `Edit` to actual name of editing entry
 - **Changed**: Viewer version from 0.4.21 to 0.5.8
+- **Changed**: `published_at` input
 - **Changed**: About page
 - **Changed**: `year` to `published_at`
 - **Changed**: Translation of year